+++ src/agents/workflows/twitter/twitterAgent.ts
@@ -1,7 +1,7 @@
 import { DynamicStructuredTool } from '@langchain/core/tools';
 import { HumanMessage } from '@langchain/core/messages';
 import { Character } from '../../../config/characters.js';
-import { TwitterApi } from '../../../services/twitter/types.js';
+import { TwitterApi } from '../../tools/twitter/types.js';
 import { createAllTwitterTools } from '../../tools/twitter/index.js';
 import { createOrchestratorRunner } from '../orchestrator/orchestratorWorkflow.js';
 import { createTwitterPrompts } from './prompts.js';
